Payroll Design changes

+++ Sales Project/MySales/MySales.DL/EmpAttendanceDL.cs
@@ -38,7 +38,7 @@ namespace MySales.DL
                                 empAttDetails.Year = year;
                                 empAttDetails.WorkDays = Convert.ToInt64(dr["WorkDays"]);
                                 empAttDetails.LeaveDays = Convert.ToInt64(dr["LeaveDays"]);
-                                empAttDetails.Overtime = Convert.ToDecimal(dr["Overtime"]);
+                                empAttDetails.Overtime = Convert.ToInt64(dr["Overtime"]);
                                 empAttDetails.CreateDate = dr["CreateDate"] == DBNull.Value ? DateTime.MinValue : Convert.ToDateTime(dr["CreateDate"]);
                                 empAttDetails.ModifiedDate = dr["ModifiedDate"] == DBNull.Value ? DateTime.MinValue : Convert.ToDateTime(dr["ModifiedDate"]);
